Move ExchangeType.cs to root namespace
