+++ Source/EasyNetQ.IntegrationTests/Advanced/When_declare_an_exchange_with_different_properties.cs
@@ -1,5 +1,4 @@
 using RabbitMQ.Client.Exceptions;
-using ExchangeType = EasyNetQ.Topology.ExchangeType;
 
 namespace EasyNetQ.IntegrationTests.Advanced;
