+++ Source/EasyNetQ.IntegrationTests/Advanced/When_publish_to_non_existent_exchange.cs
@@ -1,6 +1,5 @@
 using EasyNetQ.Topology;
 using RabbitMQ.Client.Exceptions;
-using ExchangeType = EasyNetQ.Topology.ExchangeType;
 
 namespace EasyNetQ.IntegrationTests.Advanced;
